+++ .github/workflows/maven.yml
@@ -47,8 +47,8 @@ jobs:
         run: >
           mvn --batch-mode -DskipTests sonar:sonar
           -Dsonar.host.url=https://sonarcloud.io
-          -Dsonar.organization=powsybl-ci-github
-          -Dsonar.projectKey=com.powsybl:powsybl-study-server
+          -Dsonar.organization=gridsuite
+          -Dsonar.projectKey=org.gridsuite:study-server
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
@@ -64,7 +64,7 @@ jobs:
 
       - name: Broadcast update event
         if: github.ref == 'refs/heads/master'
-        uses: geofjamg/broadcast-event@master
+        uses: gridsuite/broadcast-event@master
         with:
           token: ${{ secrets.REPO_ACCESS_TOKEN }}
           event-type: study_server_updated
