+++ README.md
@@ -1,8 +1,7 @@
-[![Actions Status](https://github.com/powsybl/powsybl-study-server/workflows/CI/badge.svg)](https://github.com/powsybl/powsybl-study-server/actions)
-[![Coverage Status](https://sonarcloud.io/api/project_badges/measure?project=com.powsybl%3Apowsybl-study-server&metric=coverage)](https://sonarcloud.io/component_measures?id=com.powsybl%3Apowsybl-study-server&metric=coverage)
+[![Actions Status](https://github.com/gridsuite/study-server/workflows/CI/badge.svg)](https://github.com/gridsuite/study-server/actions)
+[![Coverage Status](https://sonarcloud.io/api/project_badges/measure?project=org.gridsuite%3Astudy-server&metric=coverage)](https://sonarcloud.io/component_measures?id=org.gridsuite%3Astudy-server&metric=coverage)
 [![MPL-2.0 License](https://img.shields.io/badge/license-MPL_2.0-blue.svg)](https://www.mozilla.org/en-US/MPL/2.0/)
-[![Join the community on Spectrum](https://withspectrum.github.io/badge/badge.svg)](https://spectrum.chat/powsybl)
-# powsybl-study-server
+# study-server
 
 ### Run integration tests
