+++ pom.xml
@@ -10,8 +10,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.powsybl</groupId>
-    <artifactId>powsybl-study-server</artifactId>
+    <groupId>org.gridsuite</groupId>
+    <artifactId>gridsuite-study-server</artifactId>
     <packaging>jar</packaging>
     <version>1.0.0-SNAPSHOT</version>
 
@@ -23,9 +23,9 @@
     </licenses>
 
     <scm>
-        <connection>scm:git:https://github.com/powsybl/powsybl-study-server.git</connection>
-        <developerConnection>scm:git:https://github.com/powsybl/powsybl-study-server.git</developerConnection>
-        <url>https://github.com/powsybl/powsybl-study-server</url>
+        <connection>scm:git:https://github.com/gridsuite/study-server.git</connection>
+        <developerConnection>scm:git:https://github.com/gridsuite/study-server.git</developerConnection>
+        <url>https://github.com/gridsuite/study-server</url>
     </scm>
 
     <developers>
