+++ src/main/java/org/gridsuite/study/server/StudyApi.java
@@ -4,17 +4,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-
-package com.powsybl.study.server;
+package org.gridsuite.study.server;
 
 /**
- * @author Abdelsalem HEDHILI <abdelsalem.hedhili at rte-france.com>
+ * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
  */
 
-public final class CassandraConstants {
+final class StudyApi {
 
-    public static final String KEYSPACE_STUDY = "study";
-
-    private CassandraConstants() {
+    private StudyApi() {
     }
+
+    static final String API_VERSION = "v1";
 }
