+++ src/main/java/org/gridsuite/study/server/StudySwaggerConfig.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.study.server;
+package org.gridsuite.study.server;
 
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
