+++ src/main/java/org/gridsuite/study/server/dto/CaseInfos.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.study.server.dto;
+package org.gridsuite.study.server.dto;
 
 import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
@@ -12,18 +12,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-
 @AllArgsConstructor
-@NoArgsConstructor
 @Getter
-@ApiModel("Study attributes")
-public class StudyInfos {
+@NoArgsConstructor
+@ApiModel("Case infos")
+public class CaseInfos {
 
-    String studyName;
+    private String name;
 
-    String description;
-
-    String caseFormat;
+    private String format;
 }
