+++ src/main/java/org/gridsuite/study/server/dto/NetworkInfos.java
@@ -1,26 +1,28 @@
 /**
- * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.study.server.dto;
+package org.gridsuite.study.server.dto;
 
-import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.UUID;
+
 /**
- * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
  */
+
 @AllArgsConstructor
-@Getter
 @NoArgsConstructor
-@ApiModel("Case infos")
-public class CaseInfos {
+@Getter
+public class NetworkInfos {
 
-    private String name;
+    private UUID networkUuid;
 
-    private String format;
+    private String networkId;
+
 }
