+++ src/main/java/org/gridsuite/study/server/dto/StudyInfos.java
@@ -1,16 +1,15 @@
 /**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.study.server.dto;
+package org.gridsuite.study.server.dto;
 
+import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
 
 /**
  * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
@@ -19,10 +18,12 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class NetworkInfos {
+@ApiModel("Study attributes")
+public class StudyInfos {
 
-    private UUID networkUuid;
+    String studyName;
 
-    private String networkId;
+    String description;
 
+    String caseFormat;
 }
