+++ src/main/java/org/gridsuite/study/server/repository/Study.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.study.server.repository;
+package org.gridsuite.study.server.repository;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
